+++ functions.php
@@ -236,7 +236,16 @@ function disable_emojis_remove_dns_prefetch( $urls, $relation_type ) {
 return $urls;
 }
 
+/* from https://wordpress.org/support/topic/what-is-wp-json-and-api-w-org-seems-like-theyre-slowing-my-site-load/ */
+// Disable REST API link tag
+remove_action('wp_head', 'rest_output_link_wp_head', 10);
+// Disable oEmbed Discovery Links
+remove_action('wp_head', 'wp_oembed_add_discovery_links', 10);
+// Disable REST API link in HTTP headers
+remove_action('template_redirect', 'rest_output_link_header', 11, 0);
 
+/* don't add resource hints */
+remove_action('wp_head', 'wp_resource_hints', 2);
 
 /**
  * Implement the Custom Header feature.